# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Project Kessan API
  description: |
    AI-Powered Japanese Stock Trend Analysis Platform
    
    Project Kessan (決算 - "Financial Results") provides institutional-grade investment analysis 
    for Japanese stocks through AI-powered insights, real-time data, and comprehensive financial analysis.
    
    ## Features
    - **AI Analysis**: Multi-horizon stock analysis using Google Gemini API
    - **Real-time Data**: Live stock prices and market indices
    - **Japanese Focus**: Specialized for Tokyo Stock Exchange (TSE) listed companies
    - **News & Sentiment**: Japanese language news analysis with sentiment scoring
    - **Subscription Tiers**: Flexible pricing with quota management
    - **Watchlist Management**: Personal stock tracking and monitoring
    
    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limits
    - **Free Tier**: 20 requests/minute (anonymous), 100 requests/minute (authenticated)
    - **Pro Tier**: 300 requests/minute
    - **Business Tier**: 1000 requests/minute
    
    ## Error Handling
    The API returns structured error responses with appropriate HTTP status codes:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource not found)
...
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /stocks/{ticker}:
    get:
      tags:
        - Stocks
      summary: Get stock details
      description: |
        Get detailed information for a specific Japanese stock.
        Includes current price, financial metrics, and company information.
      security: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{4}$'
          description: 4-digit Japanese stock ticker (e.g., "7203" for Toyota)
      responses:
        '200':
          description: Stock details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDetail'
        '400':
          description: Invalid ticker format
        '404':
          description: Stock not found
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get current user's profile information
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithProfile'
        '401':
          description: Authentication required

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update current user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        timezone:
          type: string
          default: Asia/Tokyo
        notification_preferences:
          type: object
          additionalProperties: true
    UserWithProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/UserProfile'
    UserProfileUpdate:
      type: object
      properties:
        display_name:
          type: string
          maxLength: 50
        avatar_url:
          type: string
          format: uri
        timezone:
          type: string
        notification_preferences:
          type: object
          additionalProperties: true

    # Stock Schemas
    Stock:
      type: object
      properties:
        ticker:
          type: string
          pattern: '^[0-9]{4}$'
          description: 4-digit Japanese stock ticker
        company_name_jp:
          type: string
          description: Company name in Japanese
        company_name_en:
          type: string
          description: Company name in English
        sector_jp:
          type: string
          description: Sector in Japanese
        industry_jp:
          type: string
          description: Industry in Japanese
        description:
          type: string
          description: Company description
        is_active:
          type: boolean
          description: Whether stock is actively traded
    StockDetail:
      allOf:
        - $ref: '#/components/schemas/Stock'
        - type: object
          properties:
            current_price:
              type: number
              format: float
              description: Current stock price in JPY
            price_change:
              type: number
              format: float
              description: Price change from previous close
            price_change_percent:
              type: number
              format: float
              description: Percentage change from previous close
            volume:
              type: integer
              description: Current trading volume
            market_cap:
              type: integer
              description: Market capitalization in JPY
            pe_ratio:
              type: number
              format: float
              description: Price-to-earnings ratio
            pb_ratio:
              type: number
              format: float
              description: Price-to-book ratio
            dividend_yield:
              type: number
              format: float
              description: Dividend yield percentage
            week_52_high:
              type: number
              format: float
              description: 52-week high price
            week_52_low:
              type: number
              format: float
              description: 52-week low price
...